fix: reading_time приводится к int в BlogPost
ceil() возвращает float, из-за чего getReadingTimeAttribute() падал с TypeError на статьях с временем чтения больше 1 минуты.

+++ src/app/Models/BlogPost.php
@@ -151,7 +151,7 @@ final class BlogPost extends Model
     public function getReadingTimeAttribute(): int
     {
         $wordCount = str_word_count(strip_tags($this->content));
-        $minutes = ceil($wordCount / 200);
+        $minutes = (int) ceil($wordCount / 200);
 
         return max(1, $minutes); // Минимум 1 минута
     }
